Changed wildcard behavior.
It is now possible to request a navigation state that contains wildcards, when you are
already at a path with at least as much elements as the position of the last wildcard.

Example:

If the current state is /day/program/ and you want to navigate back to /day/, you may
pass /*/ to the Navigator and it will unmask the wildcard * to "day".

Similarly, if you are at /night/program/ and you want to get back to /night/, the same
requested path /*/ will get you there.