Added 'hide' feature to the responder types.
Any responder you want to hide at a (sub)path of where it is already used for 'show', you can now force it to hide by using the Navigator.addResponderHide() method. See source for more details.

I need to find a nicer alternative to using "debug/warn/notice/error" logs in projects like this. Throwing errors doesn't cover the need for logging stuff, and including the classes makes it hard to use as a submodule. Any tips are welcome...