New feature: Unobtrusive logging!
With the new log variable, logging is now routed through a public variable which removes the depencency of my other logging packages. Feel free to assign any new ITraceable implementation to the log variable, to enable your own logging package and read along with the wonderous internals of the navigator package!